refactor: split AuthTokenManager (shared base + JRuby subclass), centralize Map→Hash via PlainMapConverter
AuthTokenManager:
- Shared `Internal::InternalAuthTokenManager` (lib/shared/...) holds
  the Proc storage and Ruby semantics.
- JRuby's `Neo4j::Driver::AuthTokenManager` subclasses it, includes the
  Java interface in the class body (required for the Java-side proxy
  to dispatch correctly — retroactive include leaves the proxy stale),
  wraps `get_token` in a CompletionStage via super, and maps the Java
  SecurityException via super before handing to the user.
- MRI aliases `AuthTokenManager` to the same base, no Java involved.

Map→Hash:
- New `PlainMapConverter` module gives every `Java::JavaUtil::Map` an
  `as_ruby_object` method that returns a recursively-converted Ruby
  Hash with symbol keys. `MapConverter` (for Value-of-Map wrappers)
  now uses it too — `asMap(Function.identity).as_ruby_object`.
- `AuthToken#to_h` collapses to `to_map.as_ruby_object`. Output shape
  matches MRI's hash (symbol keys, only set fields, recursive value
  conversion for custom-auth `parameters`).
